HBASE-25819 Fix style issues for StochasticLoadBalancer (#3207)
