+++ hbase-server/src/main/java/org/apache/hadoop/hbase/master/balancer/RackLocalityCostFunction.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -17,10 +17,23 @@
  */
 package org.apache.hadoop.hbase.master.balancer;
 
-public class DummyCostFunction extends CostFunction {
+import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.hbase.master.balancer.BalancerClusterState.LocalityType;
+import org.apache.yetus.audience.InterfaceAudience;
+
+@InterfaceAudience.Private
+class RackLocalityCostFunction extends LocalityBasedCostFunction {
+
+  private static final String RACK_LOCALITY_COST_KEY =
+    "hbase.master.balancer.stochastic.rackLocalityCost";
+  private static final float DEFAULT_RACK_LOCALITY_COST = 15;
+
+  public RackLocalityCostFunction(Configuration conf) {
+    super(conf, LocalityType.RACK, RACK_LOCALITY_COST_KEY, DEFAULT_RACK_LOCALITY_COST);
+  }
 
   @Override
-  protected double cost() {
-    return 0;
+  int regionIndexToEntityIndex(int region) {
+    return cluster.getRackForRegion(region);
   }
 }
